CRAG-travelplanner/pyproject.toml
[project]
name = "crag-travelplanner"
version = "0.1.0"
description = "CRAG-based Travel Planning Agent with Hybrid Search"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "CRAG Team"}
]

dependencies = [
    # LangChain生态
    "langchain>=1.0.0",
    "langchain-core>=1.2.0",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.0",
    "langgraph-checkpoint>=3.0.0",
    "langsmith>=0.4.0",

    # OceanBase向量数据库
    "langchain-oceanbase>=0.2.7",
    "pyobvector>=0.2.20",
    "pymysql>=1.1.0",
    "aiomysql>=0.3.0",
    "sqlalchemy>=2.0.0",
    "sqlglot>=28.0.0",

    # LLM和工具
    "openai>=2.14.0",
    "tiktoken>=0.7.0",

    # 配置和环境
    "pydantic>=2.7.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",

    # 工具库
    "tenacity>=9.0.0",
    "requests>=2.32.0",
    "httpx>=0.28.0",

    # 数据处理
    "pandas>=2.0.0",
    "numpy>=2.0.0",

    # API服务
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",

    # 加密和安全
    "cryptography>=44.0.0",
]

[project.optional-dependencies]
dev = [
    # 测试
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",

    # 代码质量
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",

    # 类型检查
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.0",
]

# 额外LLM支持（可选）
gemini = [
    "langchain-google-genai>=2.0.0",
]

anthropic = [
    "anthropic>=0.39.0",
]

# 数据生成工具
data = [
    "faker>=25.0.0",
    "geopy>=2.4.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["seekdb_agent*"]

[tool.black]
line-length = 100
target-version = ['py311']
exclude = '''
/(
    \.git
  | \.venv
  | \.eggs
  | \.tox
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=seekdb_agent",
    "--cov-report=term-missing",
    "--cov-report=html",
]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "performance: marks tests as performance/benchmark tests",
]

[tool.coverage.run]
source = ["seekdb_agent"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
